Added: King Safety Platforms
Added a feature which allows for safety platform support. Basically making it so, when a new player becomes the king, the king will be teleported to a specified location within the map, typically a safe spot on top of a flying platform for a very short duration before the platform dissappears and the king is once more thrown in to the fight.

These platform locations can be easily defined or changed by modifying the file: addons/sourcemod/configs/KingMod/platforms.txt.