Added: Platform Administrator Command
Added the sm_platform command which will send a message to the user in their console, containing the coordinate location of the player's current position as well as the current map's name in a structured fashion where it can be easily copied and inserted in to the configs/KingMod/platforms.txt file, making the adding of platform teleportation destinations easy.

The command is restricted to only b available for administrators with root access.