Fixed: Auto-Assigned No Respawn
Fixed an issue where where the player that would be auto-assigned to a team after the mp_force_pick_time had run out would not be respawned.

In addition also added an auto-assigning system that accounts for difference in team sizes and team scores, to try and accomodate the notion of fair play through automation.