Added: Sticky Grenades King Power
Added the first power that a king can acquire which will provide the king with an infinite amount of high explosive grenades, and they will stick to walls, objects and players they hit.

Also fixed an issue in the power chooser, which would cause the new king not to receive a power.