Added: Two New King Powers
Added the carpet bombing flashbangs power to the list of obtainable powers, along with the scout no scope power.

The carpet bombing flashbang power causes the king's to be able to throw 3-5 flashbangs at once, all dealing up 1337 damage if an enemy is hit by them.

The scout no scope power is quite self-explanatory, the king is given a ssg08 but the king cannot use the zoom functionality.