Added: Sentry Gun Modifications
Changed the health of the sentry guns to 999 causing them to turn invulnerable.

Changed the damage sentry guns inflict upon players to 65% of what they normally would inflict.

Also reduced the sound produced by the sentry guns to 12% of the sound they normally would produce. Big shoutout to Backwards for helping me with the sound and damage solutions.