Added: Riot Power & Fixed Performance Bug
Added the Riot power as a new attainable power by the king. This power provides the king with a riot shield, and changes everybody's health to 1.

Also fixed a major performance issue. When changing map the server's var would climb to 65+ up to 110~ ms, because of an issue related to attempting to change the sv_cheats convar.