Added: Chuck Norris Power
Added the Chuck Norris power to the list of powers that the king can acquire. When wielding the Chuck Norris power the player will only have his fists at his disposal, but they deal a random amount of damage and have a chance to obliterate the enemy by dealing 9000 damage.