Added: Zombie Apocalypse Power
Added a new obtainable power which will turn the king in to a zombie with 325 health and reduced movement speed. The king's entire team will also turn in to zombies with 150 health, all healthshots will be removed and new ones will be unobtainable.
The sky will also change, and ambience music will start playing.