Added: Hammer Time Power & Other Power Foundations
Added the Hammer Time power to the list of obtainable powers by the king, and created the base foundation for four other powers that will be added.

When the Hammer Time power is active the king is provided a hammer, and hitting enemies will knock them down in the ground for 2.25 seconds. The hammer also deals increased damage compared to normal.