Added: Platform Toggling & Fixed Issues
Fixed issues where players would become stuck in the platforms as a result of the platforms being re-enabled briefly after the newly crowned king would leave them.

The platform toggling is now built-in to the game mode itself, to also make it easier for map creators to create maps for the King Mod game mode.